add initial load chunks by renderdistance

+++ engine/src/terrains/voxel/Voxel.h
@@ -2,7 +2,7 @@
 #include <glm/glm.hpp>
 
 namespace TerrainEngine::Voxel {
-    const glm::vec3 POSITIONS[8] = {
+    const glm::ivec3 POSITIONS[8] = {
         {0, 0, 0},
         {1, 0, 0},
         {1, 1, 0},
@@ -13,7 +13,7 @@ namespace TerrainEngine::Voxel {
         {0, 1, 1},
     };
 
-    const glm::vec3 FACES[6][4] = {
+    const glm::ivec3 FACES[6][4] = {
         {POSITIONS[0], POSITIONS[1], POSITIONS[2], POSITIONS[3]}, // front  (z-)
         {POSITIONS[5], POSITIONS[4], POSITIONS[7], POSITIONS[6]}, // back   (z+)
         {POSITIONS[4], POSITIONS[0], POSITIONS[3], POSITIONS[7]}, // left   (x-)
@@ -22,7 +22,7 @@ namespace TerrainEngine::Voxel {
         {POSITIONS[4], POSITIONS[5], POSITIONS[1], POSITIONS[0]} // bottom (y-)
     };
 
-    const glm::vec3 FACE_NORMALS[6] = {
+    const glm::ivec3 FACE_NORMALS[6] = {
         {0, 0, -1}, // front
         {0, 0, 1}, // back
         {-1, 0, 0}, // left
