+++ engine/src/terrains/voxel/VoxelChunk.h
@@ -6,18 +6,18 @@
 namespace TerrainEngine {
     class VoxelChunk {
     public:
-        VoxelChunk(size_t size, glm::vec3 position);
+        VoxelChunk(size_t size, glm::ivec3 position);
 
         ~VoxelChunk() = default;
 
         std::shared_ptr<Mesh> CreateMesh(const std::vector<VoxelChunk>& neighboringChunks) const;
 
-        glm::vec3 GetPosition() const { return position; }
+        glm::ivec3 GetPosition() const { return position; }
 
         int GetVoxelType(glm::ivec3 coord) const;
     private:
         std::vector<int> voxelMap;
-        glm::vec3 position;
+        glm::ivec3 position;
         size_t size{};
     };
 }
